+++ examples/code-http-anthropic/index.ts
@@ -35,6 +35,7 @@ new Workflow(app, "my-workflow", {
 	definition: [
 		Chain.start(
 			new Webhook("webhook", {
+				label: "Webhook Trigger",
 				parameters: {
 					responseMode: "responseNode",
 					httpMethod: "GET",
@@ -80,7 +81,6 @@ new Workflow(app, "my-workflow", {
 					}),
 					parameters: {
 						mode: "runOnceForEachItem",
-						language: "python",
 						pythonCode: PythonFunction.from({
 							projectRoot: path.join(__dirname, "python-function"),
 							entrypoint: "generate-prompt.py",